float32 instead of float64 everywhere, ~2x speedup due to SIMD

+++ optimization/strategy.go
@@ -16,7 +16,7 @@ type GameResult interface {
 	Close()
 	Zero() GameResult
 	Copy() GameResult
-	Add(other GameResult, weight float64) GameResult
+	Add(other GameResult, weight float32) GameResult
 	Max(other GameResult) GameResult
 	Shift(offset int) GameResult
 }
