+++ roll.go
@@ -120,7 +120,7 @@ func (r Roll) PossibleHolds() []Roll {
 	return holds[r]
 }
 
-func (r Roll) Probability() float64 {
+func (r Roll) Probability() float32 {
 	return probabilities[r]
 }
 
@@ -305,14 +305,14 @@ func multinomial(n int, k []int) int {
 	return result
 }
 
-func computeProbability(roll Roll) float64 {
+func computeProbability(roll Roll) float32 {
 	n := multinomial(roll.NumDice(), roll.Counts())
 	d := pow(NSides, roll.NumDice())
-	return float64(n) / float64(d)
+	return float32(n) / float32(d)
 }
 
-func computeAllProbabilities() []float64 {
-	result := make([]float64, MaxRoll)
+func computeAllProbabilities() []float32 {
+	result := make([]float32, MaxRoll)
 	for roll := Roll(0); roll < MaxRoll; roll++ {
 		if roll.NumDice() > NDice {
 			continue // Not a valid Yahtzee roll.
